feat: add support for api adapters (#770)
The current implementation of this library talks to the CMA via REST over HTTP. This PR adds `Adapter`s. These adapters add an abstraction layer that enables us to replace the existing REST communication to the server with a different technology or protocol.

An adapter can be specified via the `apiAdapter` param when calling `createClient`. All existing network request logic was moved into `RestAdapter`. If no adapter is specified, `RestAdapter` is used. Currently, there is no alternative adapter implementation.

This PR is not a breaking change! The PR should have no impact on the observable behavior of the library.